+++ src/index.ts
@@ -4,6 +4,7 @@ export * from './public/mixin';
 export * from './public/fields/scalars';
 export * from './public/fields/enums';
 export * from './public/fields/relations';
+export * from './public/fields/unsupported';
 export * as Compound from './public/fields/compounds';
 export * as Types from './types';
 export * from './public/db';