+++ src/public/db/mysql.ts
@@ -6,7 +6,7 @@ const _ = db('mysql');
 // prettier-ignore
 export default {
   Char: (value: string) =>       _.String('Char', value),
-  VarChar: (value: string) =>    _.String('VarChar', value),
+  VarChar: (value: number) =>    _.String('VarChar', value),
   TinyInt: (value: number) =>    _.Boolean('TinyInt', value),
   UnsignedBigInt:                _.BigInt('UnsignedBigInt'),
   BigInt:                        _.BigInt('BigInt'),
@@ -24,10 +24,10 @@ export default {
   Blob:                          _.Bytes('Blob'),
   Binary:                        _.Bytes('Binary'),
   Bit: (value: number) =>        _.Bytes('Bit', value),
-  Date:                          _.DateTime('Date'),
-  DateTime:                      _.DateTime('DateTime'),
-  Timestamp:                     _.DateTime('Timestamp'),
-  Time:                          _.DateTime('Time'),
+  Date: (value:number) =>        _.DateTime('Date', value),
+  DateTime: (value:number) =>    _.DateTime('DateTime', value),
+  Timestamp: (value:number) =>   _.DateTime('Timestamp', value),
+  Time: (value:number) =>        _.DateTime('Time', value),
   Json:                          _.Json('Json'),
   Text:                          _.String('Text'),
   Decimal: (...coords:Coords) => _.Decimal('Decimal', coords),